Address Travis test failures
Summary:
Due to a timezone issue, two tests in the `blog_spec` were passing
locally but failing consistently on Travis

By default, when the calling `blog.posts`, the HTTP request includes a
timestamp to limit the response to blog posts created in the last two
months. VCR recorded the timestamp in the URL and requires all future
requests to use the same timestamp. If VCR can't find a URI that matches
the request, a new, real, request is issued.

Running the tests locally will pass for anyone in the America/New_York
timezone, since this appears to be the timezone that was used when
recording the test's request. However, the tests fail on Travis because
Travis uses the `UTC` timezone.

This PR:
- Replaces `local` with `utc` to ensure consistency across timezones
- Removes assertions that focus on the API's return value in favor of
  ensuring the correct request is being sent with the expected params.
- Removes or updates tests marked as 'pending'